Update NITF schema to describe the MaskTable and parse different image compression types
Note that this does not describe the different compression types, but
does make JPEG/JFIF an element reference so that one could put a JPEG
schema on the classpath and parse the NITF JPEG data more thoroughly.
This patch includes a 'dummy' JPEG schema for testing that just treats
JPEG as a hex binary blob.