Move nil properties into dfdl:defineFormat's
The DFDL specification does not allow nil properties to be defined on a
simple type definition. So define two new named formats for the
different kinds of nillable elements and use them via dfdl:ref on
nillable elements. Also remove nil properties from the default format so
that if an element is made nillable without referencing one of the named
formats, we'll get a schema definition error. This should hopefully
remind us to use one of the existing define formats.

This does not resolve the issue where Daffodil incorrectly allows
nil properties on simple type defitionons, but does make the schema
correct according to the specification.

Fixes Issue #2