Fixed most unparsing issues and added infosets
There are still a few areas where circularity errors or other
limitations in Daffodil/DFDL are preventing "proper" unparsing (there
are many length fields that are used to determine how many times an
element occurs, but when unparsing attempting to do an outputValueCalc
of the length field by taking the dfdl:valueLength() of the array
element results in a circularity error).

There are also a set of tests that are currently commented out that are
waiting on DAFFODIL-2374 and DAFFODIL-2375 to be fixed. Once they are
fixed and the corresponding areas of the schemas are uncommented, they
should just work (although we may need to regenerate the infosets at
that point).