Change how text numbers are modeled
Text numbers are currently modeled using normal number parsing and
Daffodils pad/trim properties to add/remove leading zeros. However, for
negative numbers (i.e. types with BCS-N) this pads zeros to the left of
the negative sign instead of to the right, which is incorrect.

There is no way to have Daffodil pad to the right of a negative sign, so
this adds new simple types for text numbers with a special
textNumberPattern's that causes ICU to pad numbers as expected.
Unfortunately, the pattern is specific to the length of the string, so
we need a new type for each possible length, though this number is
relatively small..

In fact, there is actually only one signed type needed (length = 5)
since most text numbers in NITF or unsigned, but for consistency this
does the same for unsigned numbers as well, just with a simpler pattern.
And since the new types know their sizes, it means we can add
restrictions and don't have to think about if BCS_NPbig is needed or
not.

Closes #19