Fix reliance on incorrect default property resolution
According to the DFDL spec, default property resolution for the JFIF
element reference should give precedence to the JFIF default properties.
(i.e. innermost components win).

But the JFIF element assumes lengthUnits="bytes" (which is the default
for the NITF schema), but the JFIF schema defaults to
lengthUnits="bits". Since the JFIF schema default should win precedence,
we need to explicitly override this in the NITF schema. We also do this
for fillByte, which also differs in the JFIF and NITF schema.

Note that this isn't necessary on older versions of Daffodil since it
has incorrect default property precedence logic. But newer versions of
Daffodil fix this and so requires us to expilcitly override the default
properties.